added protection to routes

+++ routes/web.php
@@ -19,26 +19,26 @@ use App\Models\Listing;
 //Routes for listing
 Route::get('/', [ListingController::class, 'index']);
 
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 
 //Routes for users
-Route::get('/register', [UserConrtoller::class, 'create']);
+Route::get('/register', [UserConrtoller::class, 'create'])->middleware('guest');
 
 Route::post('/users', [UserConrtoller::class, 'store']);
 
-Route::post('/logout', [UserConrtoller::class, 'logout']);
+Route::post('/logout', [UserConrtoller::class, 'logout'])->middleware('auth');
 
-Route::get('/login', [UserConrtoller::class, 'login']);
+Route::get('/login', [UserConrtoller::class, 'login'])->name('login')->middleware('guest');
 
 Route::post('/users/authenticate', [UserConrtoller::class, 'authenticate']);
